:bug: Fix suite à la review
