+++ frontend/src/App.tsx
@@ -10,26 +10,31 @@ function App() {
                     id={1}
                     capacity={4}
                     occupied={true}
-                    hasBeenServed={false}
                     isCommandesPage={true}
-                    commandState="pending"
+                    commandState="preparing-in-kitchen"
                     commandPreparationPlace="bar"
                 />
                 <Table
                     id={2}
                     capacity={2}
                     occupied={false}
-                    hasBeenServed={false}
                     isCommandesPage={false}
                 />
                 <Table
                     id={3}
+                    capacity={2}
+                    occupied={true}
+                    isCommandesPage={true}
+                    commandState="awaiting-service"
+                    commandPreparationPlace="bar"
+                />
+                <Table
+                    id={4}
                     capacity={6}
                     occupied={true}
-                    hasBeenServed={true}
                     isCommandesPage={true}
                     commandState="served"
-                    commandPreparationPlace="kitchen"
+                    commandPreparationPlace="cuisine"
                 />
             </div>
         </div>
