+++ frontend/src/components/table/table.css
@@ -5,22 +5,22 @@
     border-radius: 0.5rem;
     width: 14rem;
     box-sizing: border-box;
-}
 
-.table-card.occupied {
-    background-color: #b62e2e;
-    color: white;
-}
+    &.occupied {
+        background-color: #b62e2e;
+        color: white;
+    }
 
-.table-card.free {
-    background-color: #77b559;
-    color: white;
-}
+    &.free {
+        background-color: #77b559;
+        color: white;
+    }
 
-.table-card h3 {
-    text-align: center;
-    font-size: 1.5rem;
-    margin-bottom: 1rem;
+    h3 {
+        text-align: center;
+        font-size: 1.5rem;
+        margin-bottom: 1rem;
+    }
 }
 
 .pay-btn {
@@ -40,13 +40,13 @@
     flex-direction: column;
     gap: 0.75rem;
     margin-top: 1rem;
-}
 
-.command-actions button {
-    font-size: 1.1rem;
-    padding: 0.5rem;
-    border-radius: 0.25rem;
-    cursor: pointer;
-    background-color: #1b6826;
-    border: solid 0.1rem #000000;
+    button {
+        font-size: 1.1rem;
+        padding: 0.5rem;
+        border-radius: 0.25rem;
+        cursor: pointer;
+        background-color: #1b6826;
+        border: solid 0.1rem #000000;
+    }
 }
